Reborn: Fix signed/unsigned and type-casting issues
Apply explicit casts and index/type fixes across multiple modules to resolve signed/unsigned mismatches and potential UB. Key changes: NAT.cpp – use UnsignedInt for node loop and adjust logging; W3DRadar.cpp/W3DDisplay.cpp/W3DScene.cpp – cast math operands to Real/Int for correct division/comparison; W3DModelDraw.cpp – cast container sizes/indices and counts to appropriate Int/UnsignedByte; WW3D::LogicFrameTimeMs – cast denominator; debug/io/profile – tighten comparisons and casts; many GeneralsMD RTS and ScriptEngine files – cast vector/list sizes to UnsignedShort/Int before serializing or indexing. Purpose: silence warnings, ensure portable/defined behavior and safer runtime arithmetic.

+++ Core/Libraries/Source/WWVegas/WW3D2/ww3d.cpp
@@ -163,7 +163,7 @@ const char* DAZZLE_INI_FILENAME="DAZZLE.INI";
 **
 ***********************************************************************************/
 
-float														WW3D::LogicFrameTimeMs = 1000.0f / WWSyncPerSecond; // initialized to something to avoid division by zero on first use
+float														WW3D::LogicFrameTimeMs = 1000.0f / static_cast<float>(WWSyncPerSecond); // initialized to something to avoid division by zero on first use
 float															WW3D::FractionalSyncMs = 0.0f;
 unsigned int											WW3D::SyncTime = 0;
 unsigned int											WW3D::PreviousSyncTime = 0;
